Refactor Figure components
